requirements.txt
psycopg2-binary==2.9.9
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
rapidfuzz==3.6.1
structlog==24.1.0
python-dotenv==1.0.1
